scripts/simple_compliance_test.py
#!/usr/bin/env python3
"""
Simple Compliance Report Test Script
Tests the basic functionality of compliance reporting without external dependencies
"""

import json
import os
import sys
from datetime import datetime
from pathlib import Path
import yaml

def load_config(config_path):
    """Load configuration file"""
    try:
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)
    except Exception as e:
        print(f"Error loading config: {e}")
        return None

def generate_mock_report(config, output_dir):
    """Generate a mock compliance report"""
    
    # Create output directory
    Path(output_dir).mkdir(parents=True, exist_ok=True)
    
    # Mock compliance data
    report_data = {
        "metadata": {
            "report_type": "Compliance Assessment",
            "generated_at": datetime.now().isoformat(),
            "reporting_period": {
                "start": "2025-07-16",
                "end": "2025-08-16"
            },
            "version": "1.0"
        },
        "executive_summary": {
            "overall_compliance_score": 95,
            "frameworks_assessed": ["SOC2", "ISO27001", "GDPR"],
            "critical_findings": 0,
            "high_findings": 1,
            "medium_findings": 3,
            "low_findings": 5
        },
        "framework_compliance": {
            "SOC2": {
                "overall_score": 96,
                "controls_assessed": 15,
                "controls_compliant": 14,
                "controls_partial": 1,
                "controls_non_compliant": 0,
                "key_controls": {
                    "CC6.1": {
                        "title": "Logical and Physical Access Controls",
                        "status": "Compliant",
                        "implementation": "Multi-factor authentication implemented, role-based access controls active",
                        "evidence": ["auth-service logs", "access control policies", "MFA configuration"]
                    },
                    "CC6.2": {
                        "title": "Authentication and Authorization", 
                        "status": "Compliant",
                        "implementation": "OAuth2/OIDC implemented with strong token security",
                        "evidence": ["OAuth endpoints", "token validation tests", "PKCE implementation"]
                    },
                    "CC7.1": {
                        "title": "System Monitoring",
                        "status": "Partial",
                        "implementation": "Comprehensive monitoring implemented, alerting configured",
                        "evidence": ["Prometheus metrics", "Elasticsearch logs", "alert configurations"],
                        "gaps": ["Need to enhance incident response automation"]
                    }
                }
            },
            "ISO27001": {
                "overall_score": 94,
                "controls_assessed": 12,
                "controls_compliant": 11,
                "controls_partial": 1,
                "controls_non_compliant": 0,
                "key_controls": {
                    "A.9.1.1": {
                        "title": "Access Control Policy",
                        "status": "Compliant", 
                        "implementation": "Comprehensive access control policies documented and implemented"
                    },
                    "A.12.4.1": {
                        "title": "Event Logging",
                        "status": "Compliant",
                        "implementation": "Audit logging with 7-year retention policy implemented"
                    }
                }
            },
            "GDPR": {
                "overall_score": 95,
                "articles_assessed": 8,
                "articles_compliant": 8,
                "articles_partial": 0,
                "articles_non_compliant": 0,
                "key_articles": {
                    "Article25": {
                        "title": "Data Protection by Design and by Default",
                        "status": "Compliant",
                        "implementation": "Privacy controls built into auth service architecture"
                    },
                    "Article32": {
                        "title": "Security of Processing",
                        "status": "Compliant", 
                        "implementation": "Encryption, access controls, and monitoring implemented"
                    }
                }
            }
        },
        "security_metrics": {
            "authentication": {
                "total_auth_attempts": 125430,
                "successful_authentications": 123876,
                "auth_success_rate": 98.76,
                "failed_authentications": 1554,
                "auth_failure_rate": 1.24,
                "mfa_usage_rate": 95.2
            },
            "system_performance": {
                "availability": 99.95,
                "avg_response_time_ms": 45,
                "p95_response_time_ms": 89,
                "error_rate": 0.12
            },
            "security_events": {
                "total_security_alerts": 15,
                "critical_alerts": 0,
                "high_alerts": 2,
                "medium_alerts": 6,
                "low_alerts": 7,
                "mean_time_to_resolution_hours": 1.5
            }
        },
        "risk_assessment": {
            "overall_risk_score": "LOW",
            "risk_categories": {
                "authentication_risks": "LOW",
                "authorization_risks": "LOW", 
                "data_protection_risks": "LOW",
                "system_security_risks": "MEDIUM",
                "compliance_risks": "LOW"
            },
            "top_risks": [
                {
                    "id": "RISK-001",
                    "title": "Dependency Vulnerabilities",
                    "severity": "MEDIUM",
                    "probability": "POSSIBLE",
                    "impact": "MEDIUM",
                    "mitigation": "Regular dependency scanning and updates implemented"
                }
            ]
        },
        "audit_findings": [
            {
                "finding_id": "AUD-001",
                "severity": "HIGH",
                "category": "Access Control",
                "title": "Privileged Account Review Process",
                "description": "Enhance automated review process for privileged accounts",
                "recommendation": "Implement quarterly automated privileged account reviews",
                "status": "OPEN",
                "target_resolution": "2025-09-15"
            },
            {
                "finding_id": "AUD-002", 
                "severity": "MEDIUM",
                "category": "Monitoring",
                "title": "Alert Response Automation",
                "description": "Enhance automated response for certain alert types",
                "recommendation": "Implement automated remediation for low-severity alerts",
                "status": "IN_PROGRESS",
                "target_resolution": "2025-08-30"
            }
        ],
        "evidence_collected": {
            "configuration_files": 25,
            "test_results": 127,
            "security_logs_analyzed": 1500000,
            "metrics_data_points": 850000,
            "policy_documents": 12
        }
    }
    
    # Write JSON report
    json_file = Path(output_dir) / f"compliance_report_{datetime.now().strftime('%Y%m%d_%H%M%S')}.json"
    with open(json_file, 'w') as f:
        json.dump(report_data, f, indent=2)
    
    # Write summary text report
    summary_file = Path(output_dir) / f"compliance_summary_{datetime.now().strftime('%Y%m%d_%H%M%S')}.txt"
    with open(summary_file, 'w') as f:
        f.write("RUST SECURITY WORKSPACE - COMPLIANCE REPORT SUMMARY\n")
        f.write("=" * 55 + "\n\n")
        f.write(f"Generated: {report_data['metadata']['generated_at']}\n")
        f.write(f"Reporting Period: {report_data['metadata']['reporting_period']['start']} to {report_data['metadata']['reporting_period']['end']}\n\n")
        
        f.write("EXECUTIVE SUMMARY\n")
        f.write("-" * 17 + "\n")
        f.write(f"Overall Compliance Score: {report_data['executive_summary']['overall_compliance_score']}%\n")
        f.write(f"Frameworks Assessed: {', '.join(report_data['executive_summary']['frameworks_assessed'])}\n")
        f.write(f"Critical Findings: {report_data['executive_summary']['critical_findings']}\n")
        f.write(f"High Findings: {report_data['executive_summary']['high_findings']}\n")
        f.write(f"Medium Findings: {report_data['executive_summary']['medium_findings']}\n")
        f.write(f"Low Findings: {report_data['executive_summary']['low_findings']}\n\n")
        
        f.write("FRAMEWORK SCORES\n")
        f.write("-" * 16 + "\n")
        for framework, data in report_data['framework_compliance'].items():
            f.write(f"{framework}: {data['overall_score']}%\n")
        
        f.write("\nKEY METRICS\n")
        f.write("-" * 11 + "\n")
        auth_metrics = report_data['security_metrics']['authentication']
        f.write(f"Authentication Success Rate: {auth_metrics['auth_success_rate']}%\n")
        f.write(f"MFA Usage Rate: {auth_metrics['mfa_usage_rate']}%\n")
        
        perf_metrics = report_data['security_metrics']['system_performance']
        f.write(f"System Availability: {perf_metrics['availability']}%\n")
        f.write(f"Average Response Time: {perf_metrics['avg_response_time_ms']}ms\n")
        
        f.write(f"\nSecurity Alerts (30 days): {report_data['security_metrics']['security_events']['total_security_alerts']}\n")
        f.write(f"Mean Time to Resolution: {report_data['security_metrics']['security_events']['mean_time_to_resolution_hours']} hours\n\n")
        
        f.write("RECOMMENDATIONS\n") 
        f.write("-" * 15 + "\n")
        for finding in report_data['audit_findings']:
            f.write(f"• {finding['title']} ({finding['severity']})\n")
            f.write(f"  {finding['recommendation']}\n")
            f.write(f"  Target: {finding['target_resolution']}\n\n")
    
    return json_file, summary_file

def main():
    # Configuration
    config_path = "config/compliance_config.yaml"
    output_dir = "reports/compliance"
    
    print("🔐 Simple Compliance Report Generator")
    print("====================================")
    
    # Load configuration
    config = load_config(config_path)
    if not config:
        print("❌ Failed to load configuration")
        return 1
    
    print("✅ Configuration loaded successfully")
    
    # Generate mock compliance report
    try:
        json_file, summary_file = generate_mock_report(config, output_dir)
        
        print(f"✅ Compliance report generated successfully!")
        print(f"📄 JSON Report: {json_file}")
        print(f"📄 Summary Report: {summary_file}")
        print(f"📁 Output Directory: {output_dir}")
        
        # Display summary
        print("\n📊 COMPLIANCE SUMMARY")
        print("=" * 21)
        with open(summary_file, 'r') as f:
            lines = f.readlines()
            # Print first 20 lines of summary
            for line in lines[:20]:
                print(line.rstrip())
        
        return 0
        
    except Exception as e:
        print(f"❌ Error generating report: {e}")
        return 1

if __name__ == "__main__":
    exit_code = main()
    sys.exit(exit_code)